# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/rankings": {
      "post": {
        "operationId": "RankingsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRankingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ranking já existia — devolvido sem criar outro.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankingViewDto"
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankingViewDto"
                }
              }
            }
          }
        },
        "tags": [
          "rankings"
        ]
      }
    },
    "/v1/rankings/me/{albumId}": {
      "get": {
        "operationId": "RankingsController_getMine",
        "parameters": [
          {
            "name": "albumId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankingViewDto"
                }
              }
            }
          }
        },
        "tags": [
          "rankings"
        ]
      }
    },
    "/v1/rankings/users/{userId}/albums/{albumId}": {
      "get": {
        "operationId": "RankingsController_getByUserAndAlbum",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "albumId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankingViewDto"
                }
              }
            }
          }
        },
        "tags": [
          "rankings"
        ]
      }
    },
    "/v1/rankings/{rankingId}": {
      "get": {
        "operationId": "RankingsController_getById",
        "parameters": [
          {
            "name": "rankingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankingViewDto"
                }
              }
            }
          }
        },
        "tags": [
          "rankings"
        ]
      },
      "delete": {
        "operationId": "RankingsController_remove",
        "parameters": [
          {
            "name": "rankingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "rankings"
        ]
      }
    },
    "/v1/rankings/{rankingId}/tracks/{trackId}": {
      "patch": {
        "operationId": "RankingsController_rate",
        "parameters": [
          {
            "name": "rankingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trackId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RateTrackDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankingViewDto"
                }
              }
            }
          }
        },
        "tags": [
          "rankings"
        ]
      }
    },
    "/v1/rankings/{rankingId}/tracks/{trackId}/ignore": {
      "patch": {
        "operationId": "RankingsController_ignore",
        "parameters": [
          {
            "name": "rankingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trackId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetTrackIgnoredDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankingViewDto"
                }
              }
            }
          }
        },
        "tags": [
          "rankings"
        ]
      }
    },
    "/v1/rankings/{rankingId}/review": {
      "patch": {
        "operationId": "RankingsController_review",
        "parameters": [
          {
            "name": "rankingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveReviewDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankingViewDto"
                }
              }
            }
          }
        },
        "tags": [
          "rankings"
        ]
      }
    },
    "/v1/rankings/{rankingId}/reset": {
      "post": {
        "operationId": "RankingsController_reset",
        "parameters": [
          {
            "name": "rankingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankingViewDto"
                }
              }
            }
          }
        },
        "tags": [
          "rankings"
        ]
      }
    },
...
  },
  "info": {
    "title": "Track by Track API",
    "description": "Contrato da API — fonte de tipos para o frontend (ver seção 2.2 do CLAUDE.md)",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateRankingDto": {
        "type": "object",
        "properties": {
          "albumId": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "albumId"
        ]
      },
      "RankingViewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "albumId": {
            "type": "string"
          },
          "entries": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "trackId": {
                  "type": "string"
                },
                "score": {
                  "type": "number"
                },
                "position": {
                  "type": "number"
                },
                "ignored": {
                  "type": "boolean"
                }
              },
              "required": [
                "trackId",
                "score",
                "position",
                "ignored"
              ]
            }
          },
          "averageScore": {
            "type": "number"
          },
          "progress": {
            "type": "object",
            "properties": {
              "rated": {
                "type": "number"
              },
              "total": {
                "type": "number"
              },
              "ignored": {
                "type": "number"
              },
              "percentage": {
                "type": "number"
              }
            },
            "required": [
              "rated",
              "total",
              "ignored",
              "percentage"
            ]
          },
          "review": {
            "type": "object",
            "properties": {
              "text": {
                "type": "string"
              },
              "favoriteTrackId": {
                "type": "string"
              },
              "worstTrackId": {
                "type": "string"
              }
            }
          },
          "isFirstCompletionBadgeActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId",
          "albumId",
          "entries",
          "averageScore",
          "progress",
          "review",
          "isFirstCompletionBadgeActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "RateTrackDto": {
        "type": "object",
        "properties": {
          "score": {
            "type": "number",
            "minimum": 0,
            "maximum": 5,
            "multipleOf": 0.5
          }
        },
        "required": [
          "score"
        ]
      },
      "SetTrackIgnoredDto": {
        "type": "object",
        "properties": {
          "ignored": {
            "type": "boolean"
          }
        },
        "required": [
          "ignored"
        ]
      },
      "SaveReviewDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 2000,
            "nullable": true
          },
          "favoriteTrackId": {
            "type": "string",
            "nullable": true
          },
          "worstTrackId": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  }
}